Added custom UI components for NativeBase badges. Added Home page. Added Loading component
Added a Home page component with the needed custom badge components containing a NativeBase badge and message text. The BadgeCustom component is dynamic. By passing in a prop of 'success' the badge can either appear as green with a checkmark icon or red with a times icon. A Loading page component has also been added but needs updates. A UNHBanner component has been added as well for use across multiple pages.

Expo was returning a missing font family error. This seems to have been fixed using an async function which loads the fonts at runtime within the App component. expo-fonts have been installed for this purpose.