Added VoiceInput component for microphone voice input testing
NOTE: Voice input currently only works on iOS devices. Android devices do not pick up on audio input.

The voice input update seeks to implement a speech-to-text process using the microphone of a mobile device. This will later be upgraded for wireless input via Bluetooth. The VoiceInput component utilizes two Expo APIs, speech and permissions. With these together, React Native is able to grant access for the use of a microphone on a mobile device and record audio files. These audio files are saved as .wav files on ioS and are immediately sent to the Google Cloud Speech API where I've implemented a Google Cloud Function which converts the .wav speech files to text. Once completed, the text is received by the mobile device and sent as an intent to IBM Watson Assistant. Watson then sends a response depending on the intent. The component is currently able to send and receive this data as intended. Next steps will include connecting to a Bluetooth microphone for wireless input and converting Watson's responses using text-to-speech.