Update VoiceInput component to work with button press rather than hold
Updated all variables, function names, and strings to use 'Charlie' rather than 'Watson'. Updated the push-to-talk button to function by pressing once to speak and again to end. This is a change from the original way of speaking which required the user to hold the button down to speak and let go to indicate completion. Pressing the button once and again seems a bit more efficient