Update VoiceInput component for look and feel. There is still a major issue with response audio playback
Updated VoiceInput UI with a slightly different looking interface. The button for speaking is now at the bottom of the screen and is placed horizontally in the center with a microphone icon. The component also contains dynamic text showing the 'Charlie Assistant' title as well as the intent and response text that is received.

There seems to be an issue with playing the same response audio file for back-to-back questions. This could potentially be resolved by successfully deleting the charlieres mp3 file, if it exists, before playback. Unfortunately, I've been unable to delete the file before-or-following audio playback without getting error messages at some point. The errors refer to a file path not being found following deletion